refactor(scheduler): bins schedule themselves; SchedulerRunner has zero op references
Bigger structural turn off #626 review. The runner is now a registration
map only; OFD lives entirely under operations/ofd/ and shows up as a
BinPacker bean. Adding a future operation type is one new @Component;
the scheduler module is untouched.

What changed:

- binpack/Bin → interface with `void schedule()`. Bins own their own
  scheduling (claim CAS, narrow to claimed, launch, mark
  SCHEDULED/PENDING).
- binpack/BinPacker → orchestration interface: `getOperationType()` plus
  `prepare(db, tableName) → List<Bin>`. Per-op-type orchestrator.
- binpack/FirstFitDecreasingBinPacker → standalone algorithm class (not
  a BinPacker in this vocabulary). Returns flat groupings
  `List<List<BinItem>>` that the per-op-type packer wraps into its own
  Bin impl. Internal `PackingBin` helper holds items + totalWeight
  during the fold so the public `Bin` interface stays minimal.
- scheduler/SchedulerRunner → rewritten as a thin dispatcher.
  Constructor takes `List<BinPacker>` via Spring injection; builds an
  immutable `Map<OperationTypeDto, BinPacker>` with `Map.copyOf`.
  `schedule(type)` calls `bp.prepare(...).forEach(Bin::schedule)`.
  Imports only `binpack.BinPacker`, `binpack.Bin`, and the model enum —
  no operations.* anywhere. Grep-able invariant.
- scheduler/config/SchedulerConfig → stripped to the shared infra
  (`WebClient`, `JobsServiceClient`, cluster id). No more OFD @Value
  fields, no more binPackers @Bean — those moved to OfdBinPacker's ctor.
- operations/ofd/OfdBinPacker → new @Component implementing BinPacker.
  Holds @Value-bound caps, an FFD instance, repos, jobs client, results
  endpoint. `prepare(...)` does load PENDING (OFD-filtered) → dedup →
  stats lookup → project to OfdBinItem → FFD pack → wrap each grouping
  in a new OfdBin.
- operations/ofd/OfdBin → new class implementing Bin. Holds the bin's
  OfdBinItems plus refs to repo/jobs client/endpoint. `@Transactional
  schedule()` does the claim CAS, partial-claim narrow, launch, mark
  SCHEDULED/PENDING. The OFD-specific job name and arg shape live here.
- operations/ofd/OfdBinItem → `currentFileCount` wraps nulls in
  Optional locally (`Optional.ofNullable(stats).map(...).map(...)
  .orElse(0L)`). DTO getters stay nullable; converting them to return
  Optional is deferred to a follow-up PR per the null-is-a-code-smell
  lesson.
- schedulerapp/SchedulerApplication → injects SchedulerRunner only;
  loops via `runner.getRegisteredOperationTypes()`.

Tests reorganized along the same lines:
- FirstFitDecreasingBinPackerTest asserts groupings (`List<List<BinItem>>`)
  instead of bins. zero-cap-disables test gone — caps are required
  positive now. maxItemsCap test uses a real positive weight cap.
- SchedulerRunnerTest slim: three tests for the dispatcher (unknown type
  throws, delegates to the right packer + schedules each returned bin,
  passes scope args through). Mock BinPacker, mock Bin, mock the
  delegation invariant.
- OfdBinPackerTest new: covers load-PENDING + dedup + stats-filter +
  groupings-to-OfdBins. Direct-constructor (no Spring).
- OfdBinTest new: covers OfdBin.schedule()'s claim + narrow + launch +
  mark paths (success, launch fails reverts to PENDING, rows already
  claimed skips, partial claim launches only the claimed subset).

Doc cleanup pass per "describe what code does, not what it doesn't do"
lesson: dropped the "no defaults" rationale and "0 disables a dimension"
sentinel from FFD's javadoc; Bin.fits (now a private PackingBin method
inside FFD) lost its `> 0` guards. Class doc reads positively now.

Global side: created ~/.claude/code-lessons.md with the null/Optional
lesson, the comment-style lesson, the no-abstract-knobs lesson, the
generics-vs-interface lesson, and the scheduler-pluggability principle.
Linked from ~/.claude/CLAUDE.md so future sessions surface them
proactively.
