+++ services/optimizer/scheduler/src/main/java/com/linkedin/openhouse/optimizer/scheduler/binpack/FirstFitDecreasingBinPacker.java
@@ -7,65 +7,57 @@ import lombok.Builder;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * First-fit-decreasing bin packer with two independent caps:
+ * First-fit-decreasing packing algorithm. Sorts items by weight descending and places each into the
+ * first group whose running totals stay at or below {@code maxWeightPerBin} and {@code
+ * maxItemsPerBin}. An item that exceeds the weight cap on its own goes into a group by itself.
  *
- * <ul>
- *   <li>{@code maxWeightPerBin} — total {@link BinItem#getWeight()} per bin
- *   <li>{@code maxItemsPerBin} — number of items per bin
- * </ul>
- *
- * <p>Both caps are explicit on construction. Neither has a default — "weight" has no domain meaning
- * at this layer, so picking a constant here would be an arbitrary knob; callers (e.g. {@link
- * com.linkedin.openhouse.optimizer.scheduler.config.SchedulerConfig}) supply the per-op- type cap
- * with the unit attached and a justification at the config site. Pass {@code 0} or a negative value
- * for either cap to disable that dimension.
- *
- * <p>An item that exceeds the weight cap on its own is placed into a bin by itself rather than
- * dropped — the scheduler never silently skips maintenance work for an oversized table.
- *
- * <p>The pack body is one stream pipeline: sort decreasing by weight, then fold each item into the
- * running list of bins via {@code Stream.collect(Supplier, BiConsumer, BiConsumer)} — the idiomatic
- * shape for an FFD-style stateful collect.
+ * <p>Returns flat groupings ({@code List<List<BinItem>>}). Callers wrap each grouping into the
+ * {@link Bin} implementation they need for their operation type.
  */
 @Slf4j
 @Builder
-public class FirstFitDecreasingBinPacker implements BinPacker {
+public class FirstFitDecreasingBinPacker {
 
   private final long maxWeightPerBin;
   private final int maxItemsPerBin;
 
-  @Override
-  public List<Bin> pack(List<BinItem> items) {
+  public List<List<BinItem>> pack(List<BinItem> items) {
     if (items == null || items.isEmpty()) {
       return new ArrayList<>();
     }
-    List<Bin> bins =
+    List<PackingBin> bins =
         items.stream()
             .sorted(Comparator.comparingLong(BinItem::getWeight).reversed())
             .collect(ArrayList::new, this::placeItem, List::addAll);
-    log.info("Packed {} items into {} bins", items.size(), bins.size());
-    return bins;
+    log.info("Packed {} items into {} groupings", items.size(), bins.size());
+    return bins.stream().map(b -> b.items).collect(java.util.stream.Collectors.toList());
   }
 
-  /**
-   * Place {@code item} into the first bin that can hold it; if none, open a fresh bin. Mutates
-   * {@code bins} — used as the accumulator step of the {@code pack} fold.
-   */
-  private void placeItem(List<Bin> bins, BinItem item) {
+  private void placeItem(List<PackingBin> bins, BinItem item) {
     bins.stream()
         .filter(b -> b.fits(item, maxWeightPerBin, maxItemsPerBin))
         .findFirst()
         .ifPresentOrElse(
             b -> b.add(item),
             () -> {
-              Bin fresh = new Bin();
-              if (!fresh.fits(item, maxWeightPerBin, maxItemsPerBin)) {
-                log.warn(
-                    "Item exceeds per-bin caps on its own; placing in dedicated bin: weight={}",
-                    item.getWeight());
-              }
+              PackingBin fresh = new PackingBin();
               fresh.add(item);
               bins.add(fresh);
             });
+  }
+
+  /** Per-bin running-totals helper used during the fold. Hidden from callers. */
+  private static class PackingBin {
+    final List<BinItem> items = new ArrayList<>();
+    long totalWeight;
+
+    boolean fits(BinItem item, long maxWeight, int maxItems) {
+      return items.size() < maxItems && totalWeight + item.getWeight() <= maxWeight;
+    }
+
+    void add(BinItem item) {
+      items.add(item);
+      totalWeight += item.getWeight();
+    }
   }
 }
