+++ services/optimizer/scheduler/src/test/java/com/linkedin/openhouse/optimizer/scheduler/binpack/FirstFitDecreasingBinPackerTest.java
@@ -22,60 +22,78 @@ class FirstFitDecreasingBinPackerTest {
   }
 
   @Test
-  void emptyInput_returnsEmptyBins() {
-    FirstFitDecreasingBinPacker packer = FirstFitDecreasingBinPacker.builder().build();
+  void emptyInput_returnsEmptyGroupings() {
+    FirstFitDecreasingBinPacker packer =
+        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(100L).maxItemsPerBin(10).build();
     assertThat(packer.pack(List.of())).isEmpty();
   }
 
   @Test
-  void singleItem_oneBin() {
+  void singleItem_oneGrouping() {
     FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000_000L).build();
-    List<Bin> bins = packer.pack(List.of(item("a", 100L)));
-    assertThat(bins).hasSize(1);
-    assertThat(bins.get(0).size()).isEqualTo(1);
+        FirstFitDecreasingBinPacker.builder()
+            .maxWeightPerBin(1_000_000L)
+            .maxItemsPerBin(10)
+            .build();
+    List<List<BinItem>> groupings = packer.pack(List.of(item("a", 100L)));
+    assertThat(groupings).hasSize(1);
+    assertThat(groupings.get(0)).hasSize(1);
   }
 
   @Test
-  void underWeightLimit_oneBin() {
+  void underWeightLimit_oneGrouping() {
     FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000_000L).build();
-    List<Bin> bins =
+        FirstFitDecreasingBinPacker.builder()
+            .maxWeightPerBin(1_000_000L)
+            .maxItemsPerBin(10)
+            .build();
+    List<List<BinItem>> groupings =
         packer.pack(List.of(item("a", 300_000L), item("b", 300_000L), item("c", 300_000L)));
-    assertThat(bins).hasSize(1);
-    assertThat(bins.get(0).size()).isEqualTo(3);
-    assertThat(bins.get(0).getTotalWeight()).isEqualTo(900_000L);
+    assertThat(groupings).hasSize(1);
+    assertThat(groupings.get(0)).hasSize(3);
+    long total = groupings.get(0).stream().mapToLong(BinItem::getWeight).sum();
+    assertThat(total).isEqualTo(900_000L);
   }
 
   @Test
-  void overWeightLimit_twoBins() {
+  void overWeightLimit_twoGroupings() {
     FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000_000L).build();
-    List<Bin> bins =
+        FirstFitDecreasingBinPacker.builder()
+            .maxWeightPerBin(1_000_000L)
+            .maxItemsPerBin(10)
+            .build();
+    List<List<BinItem>> groupings =
         packer.pack(List.of(item("a", 600_000L), item("b", 600_000L), item("c", 400_000L)));
-    assertThat(bins).hasSize(2);
-    // FFD: sort desc → 600, 600, 400. Place 600 → bin0; next 600 doesn't fit bin0, → bin1;
-    // 400 fits bin0 (total 1_000_000).
-    assertThat(bins.get(0).getTotalWeight()).isEqualTo(1_000_000L);
-    assertThat(bins.get(1).getTotalWeight()).isEqualTo(600_000L);
+    assertThat(groupings).hasSize(2);
+    // FFD: sort desc → 600, 600, 400. Place 600 → group0; next 600 doesn't fit group0 → group1;
+    // 400 fits group0 (total 1_000_000).
+    long g0Total = groupings.get(0).stream().mapToLong(BinItem::getWeight).sum();
+    long g1Total = groupings.get(1).stream().mapToLong(BinItem::getWeight).sum();
+    assertThat(g0Total).isEqualTo(1_000_000L);
+    assertThat(g1Total).isEqualTo(600_000L);
   }
 
   @Test
-  void itemLargerThanCap_getsOwnBin() {
+  void itemLargerThanCap_getsOwnGrouping() {
     FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000L).build();
-    List<Bin> bins = packer.pack(List.of(item("big", 5_000L)));
-    assertThat(bins).hasSize(1);
-    assertThat(bins.get(0).size()).isEqualTo(1);
+        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000L).maxItemsPerBin(10).build();
+    List<List<BinItem>> groupings = packer.pack(List.of(item("big", 5_000L)));
+    assertThat(groupings).hasSize(1);
+    assertThat(groupings.get(0)).hasSize(1);
   }
 
   @Test
   void sortedDescending_largestFirst() {
-    FirstFitDecreasingBinPacker packer = FirstFitDecreasingBinPacker.builder().build();
-    List<Bin> bins = packer.pack(List.of(item("small", 100L), item("large", 900_000L)));
-    assertThat(bins).hasSize(1);
+    FirstFitDecreasingBinPacker packer =
+        FirstFitDecreasingBinPacker.builder()
+            .maxWeightPerBin(2_000_000L)
+            .maxItemsPerBin(10)
+            .build();
+    List<List<BinItem>> groupings =
+        packer.pack(List.of(item("small", 100L), item("large", 900_000L)));
+    assertThat(groupings).hasSize(1);
     List<String> ids =
-        bins.get(0).items().stream()
+        groupings.get(0).stream()
             .map(TestItem.class::cast)
             .map(TestItem::getId)
             .collect(Collectors.toList());
@@ -83,24 +101,13 @@ class FirstFitDecreasingBinPackerTest {
   }
 
   @Test
-  void maxItemsCap_splitsBins() {
+  void maxItemsCap_splitsGroupings() {
     FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(0L).maxItemsPerBin(2).build();
-    List<Bin> bins =
+        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(1_000_000L).maxItemsPerBin(2).build();
+    List<List<BinItem>> groupings =
         packer.pack(List.of(item("a", 1L), item("b", 1L), item("c", 1L), item("d", 1L)));
-    assertThat(bins).hasSize(2);
-    assertThat(bins.get(0).size()).isEqualTo(2);
-    assertThat(bins.get(1).size()).isEqualTo(2);
-  }
-
-  @Test
-  void zeroCap_disablesDimension() {
-    // All caps zero → everything in one bin regardless of weight.
-    FirstFitDecreasingBinPacker packer =
-        FirstFitDecreasingBinPacker.builder().maxWeightPerBin(0L).maxItemsPerBin(0).build();
-    List<Bin> bins =
-        packer.pack(List.of(item("a", Long.MAX_VALUE / 4), item("b", Long.MAX_VALUE / 4)));
-    assertThat(bins).hasSize(1);
-    assertThat(bins.get(0).size()).isEqualTo(2);
+    assertThat(groupings).hasSize(2);
+    assertThat(groupings.get(0)).hasSize(2);
+    assertThat(groupings.get(1)).hasSize(2);
   }
 }
